Generalizing the login overlay logic
* Created a separate reusable component for this logic
* That can be integrated with other components
