+++ src/components/Common/BasicTabs.tsx
@@ -1,15 +1,14 @@
-import { Button, Grid } from "@mui/material";
 import Box from "@mui/material/Box";
 import Tab from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
 import Typography from "@mui/material/Typography";
 import * as React from "react";
+import LoginOverlaySmall from "./LoginOverlaySmall";
 
 interface Props {
   inputeditor: React.ReactNode;
   outputeditor: React.ReactNode;
   isInterectable: boolean;
-  handleLogin: () => void;
 }
 
 interface TabPanelProps {
@@ -49,7 +48,6 @@ export default function BasicTabs({
   inputeditor,
   outputeditor,
   isInterectable,
-  handleLogin,
 }: Props) {
   const [value, setValue] = React.useState(0);
 
@@ -70,68 +68,7 @@ export default function BasicTabs({
         </Tabs>
       </Box>
       <CustomTabPanel value={value} index={0}>
-        {!isInterectable && (
-          <Box
-            sx={{
-              position: "absolute",
-              bgcolor: "rgba(0, 0, 0, 0.50)",
-              color: "common.white",
-              height: "calc(100vh - 197px)",
-              width: { xs: "80%", sm: "88.5%" },
-              zIndex: 1,
-            }}
-            marginTop={4.6}
-            alignItems="center"
-            justifyContent="center"
-          >
-            <Grid
-              container
-              alignItems="center"
-              justifyContent="center"
-              height="calc(100vh - 197px)"
-            >
-              <Box
-                alignItems="center"
-                justifyContent="center"
-                flexDirection="column"
-                display="flex"
-                bgcolor="background.paper"
-                width="80%"
-                padding={1}
-                borderRadius={1}
-              >
-                <Typography
-                  variant="h6"
-                  marginBottom={2}
-                  color="common.black"
-                  textAlign="center"
-                >
-                  Please sign in to try out the Healthcare tool
-                </Typography>
-                <Button
-                  variant="contained"
-                  size="large"
-                  onClick={handleLogin}
-                  sx={{
-                    backgroundColor: "primary.main",
-                    color: "primary.contrastText",
-                    borderRadius: "8px",
-                    fontWeight: 600,
-                    fontSize: { xs: "1rem" },
-                    textTransform: "none",
-                    alignSelf: "center",
-                    "&:hover": {
-                      backgroundColor: "secondary.main",
-                    },
-                  }}
-                >
-                  Sign In
-                </Button>
-              </Box>
-              <Box alignItems="center" justifyContent="center"></Box>
-            </Grid>
-          </Box>
-        )}
+        {!isInterectable && <LoginOverlaySmall />}
         {inputeditor}
       </CustomTabPanel>
       <CustomTabPanel value={value} index={1}>
